requirements.txt
python-dotenv
pandas
numpy
certifi
pymongo[srv]>=4.0.0
scikit-learn
dill
pyaml
mlflow
dagshub
fastapi
uvicorn

# to be enabled for final package
# -e.